# openapi.yaml
openapi: 3.1.0
info:
  title: Restaurant API
  description: API for restaurant menu
  version: 1.0.0
paths:
  /api/v1/menus/:
    get:
      tags:
        - Menu
      summary: Read Menus
      operationId: read_menus_api_v1_menus__get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/MenuWithDetailsScheme'
                type: array
                title: Response Read Menus Api V1 Menus  Get
    post:
      tags:
        - Menu
      summary: Create Menu
      operationId: create_menu_api_v1_menus__post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MenuSchemeCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MenuScheme'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/menus/{menu_id}:
    get:
      tags:
        - Menu
      summary: Read Menu
      operationId: read_menu_api_v1_menus__menu_id__get
      parameters:
        - name: menu_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            title: Menu Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MenuWithDetailsScheme'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
        - Menu
      summary: Update Menu
      operationId: update_menu_api_v1_menus__menu_id__patch
      parameters:
        - name: menu_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            title: Menu Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MenuSchemeCreate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MenuScheme'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
        - Menu
      summary: Delete Menu
      operationId: delete_menu_api_v1_menus__menu_id__delete
      parameters:
        - name: menu_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            title: Menu Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MenuScheme'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/menus/{menu_id}/submenus:
    post:
      tags:
        - Submenu
      summary: Create Submenu
      operationId: create_submenu_api_v1_menus__menu_id__submenus_post
      parameters:
        - name: menu_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            title: Menu Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmenuSchemeCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubmenuScheme'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
        - Submenu
      summary: Read Submenus
      operationId: read_submenus_api_v1_menus__menu_id__submenus_get
      parameters:
        - name: menu_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            title: Menu Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SubmenuWithDishCountScheme'
                title: Response Read Submenus Api V1 Menus  Menu Id  Submenus Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/menus/{menu_id}/submenus/{submenu_id}:
    get:
      tags:
        - Submenu
      summary: Read Submenu
      operationId: read_submenu_api_v1_menus__menu_id__submenus__submenu_id__get
      parameters:
        - name: menu_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            title: Menu Id
        - name: submenu_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            title: Submenu Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubmenuWithDishCountScheme'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
        - Submenu
      summary: Update Submenu
      operationId: update_submenu_api_v1_menus__menu_id__submenus__submenu_id__patch
      parameters:
        - name: menu_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            title: Menu Id
        - name: submenu_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            title: Submenu Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmenuSchemeCreate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubmenuScheme'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
        - Submenu
      summary: Delete Submenu
      operationId: delete_submenu_api_v1_menus__menu_id__submenus__submenu_id__delete
      parameters:
        - name: menu_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            title: Menu Id
        - name: submenu_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            title: Submenu Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubmenuScheme'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/menus/{menu_id}/submenus/{submenu_id}/dishes:
    post:
      tags:
        - Dish
      summary: Create Dish
      operationId: create_dish_api_v1_menus__menu_id__submenus__submenu_id__dishes_post
      parameters:
        - name: menu_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            title: Menu Id
        - name: submenu_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            title: Submenu Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DishSchemeCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DishScheme'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
        - Dish
      summary: Read Dishes
      operationId: read_dishes_api_v1_menus__menu_id__submenus__submenu_id__dishes_get
      parameters:
        - name: menu_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            title: Menu Id
        - name: submenu_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            title: Submenu Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DishScheme'
                title: >-
                  Response Read Dishes Api V1 Menus  Menu Id  Submenus  Submenu
                  Id  Dishes Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/menus/{menu_id}/submenus/{submenu_id}/dishes/{dish_id}:
    get:
      tags:
        - Dish
      summary: Read Dish
      operationId: >-
        read_dish_api_v1_menus__menu_id__submenus__submenu_id__dishes__dish_id__get
      parameters:
        - name: menu_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            title: Menu Id
        - name: submenu_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            title: Submenu Id
        - name: dish_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            title: Dish Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DishScheme'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
        - Dish
      summary: Update Dish
      operationId: >-
        update_dish_api_v1_menus__menu_id__submenus__submenu_id__dishes__dish_id__patch
      parameters:
        - name: menu_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            title: Menu Id
        - name: submenu_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            title: Submenu Id
        - name: dish_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            title: Dish Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DishSchemeCreate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DishScheme'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
        - Dish
      summary: Delete Dish
      operationId: >-
        delete_dish_api_v1_menus__menu_id__submenus__submenu_id__dishes__dish_id__delete
      parameters:
        - name: menu_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            title: Menu Id
        - name: submenu_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            title: Submenu Id
        - name: dish_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            title: Dish Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DishScheme'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
    DishScheme:
      properties:
        title:
          type: string
          title: Title
        description:
          type: string
          title: Description
        price:
          type: string
          title: Price
        id:
          type: string
          format: uuid
          title: Id
      type: object
      required:
        - title
        - description
        - price
        - id
      title: DishScheme
    DishSchemeCreate:
      properties:
        title:
          type: string
          title: Title
        description:
          type: string
          title: Description
        price:
          type: string
          title: Price
      type: object
      required:
        - title
        - description
        - price
      title: DishSchemeCreate
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    MenuScheme:
      properties:
        title:
          type: string
          title: Title
        description:
          type: string
          title: Description
        id:
          type: string
          format: uuid
          title: Id
      type: object
      required:
        - title
        - description
        - id
      title: MenuScheme
    MenuSchemeCreate:
      properties:
        title:
          type: string
          title: Title
        description:
          type: string
          title: Description
      type: object
      required:
        - title
        - description
      title: MenuSchemeCreate
    MenuWithDetailsScheme:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        title:
          type: string
          title: Title
        description:
          type: string
          title: Description
        submenus_count:
          type: integer
          title: Submenus Count
        dishes_count:
          type: integer
          title: Dishes Count
      type: object
      required:
        - id
        - title
        - description
        - submenus_count
        - dishes_count
      title: MenuWithDetailsScheme
    SubmenuScheme:
      properties:
        title:
          type: string
          title: Title
        description:
          type: string
          title: Description
        id:
          type: string
          format: uuid
          title: Id
      type: object
      required:
        - title
        - description
        - id
      title: SubmenuScheme
    SubmenuSchemeCreate:
      properties:
        title:
          type: string
          title: Title
        description:
          type: string
          title: Description
      type: object
      required:
        - title
        - description
      title: SubmenuSchemeCreate
    SubmenuWithDishCountScheme:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        title:
          type: string
          title: Title
        description:
          type: string
          title: Description
        dishes_count:
          type: integer
          title: Dishes Count
      type: object
      required:
        - id
        - title
        - description
        - dishes_count
      title: SubmenuWithDishCountScheme
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
              - type: string
              - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
        - loc
        - msg
        - type
      title: ValidationError